+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
 ## [Unreleased]
+* [#5](https://github.com/Blackjacx/ASCKit/pull/5): Paging - [@blackjacx](https://github.com/blackjacx).
 * [#4](https://github.com/Blackjacx/ASCKit/pull/4): Key location via "Sources" - [@blackjacx](https://github.com/blackjacx).